requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.36
jinja2>=3.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.2.1
resend>=2.0.0
httpx>=0.27.0
cryptography>=42.0.0
